Move ping classes to own file
