+++ DiscordBot/Commands/Interactive2/Ping/InsultCommandRequest.cs
@@ -1,0 +1,9 @@
+using DiscordBot.Commands.Interactive2.Base.Requests;
+using DiscordBot.Common.Models.Enums;
+
+namespace DiscordBot.Commands.Interactive2.Ping;
+
+public class InsultCommandRequest : ApplicationCommandRequestBase<InsultSubCommandDefinition> {
+	public InsultCommandRequest(ApplicationCommandContext context) : base(context) { }
+	public override AuthorizationRoles MinimumAuthorizationRole => AuthorizationRoles.ClanGuest;
+}
